T267 Remove unused code and code splitting
Removed code splitting because it didn't seem to be used. Also for the
routes we have specified, it seems that we are trying to make sure
things are loaded up front. This can be changed back of course, but at
least the documentation above the code splitting should be more
descriptive.